fspec-fromdir: Add ability to select the parts of the directory to index
Currently, fspec-fromdir can only index an entire directory, but
it may be useful to include only certain contents, for instance to
ignore .git or other unwanted files.

To support this, change the existing chdir operand to the -C option
(a convention used by a variety of other tools), and use any operands
as a list of contents under that directory to include. If no operands
are present, the entire directory is indexed.